11-07-2018

+++ lib/binary_source_20180523.cpp
@@ -168,7 +168,7 @@ bool BinarySource::runBlock(void) {
 		//std::random_device rd;
 		//std::mt19937 gen(rd()); // I think using rd to seed the mt19937 may always produce the same sequence
 		//std::uniform_int_distribution<> dis(0, 1);
-		std::default_random_engine rd(time(0));
+		std::default_random_engine rd((unsigned int)time(0));
 		std::uniform_real_distribution<> dis(0, 1);
 
 		
@@ -176,8 +176,7 @@ bool BinarySource::runBlock(void) {
 			t_binary aux;
 			if (dis(rd) < probabilityOfZero) {
 				aux = 0;
-			}
-			else {
+			} else {
 				aux = 1;
 			};
 			for (auto k = 0; k < numberOfOutputSignals; k++) {
